Lots more refactoring
