+++ identity-service/src/notifications/constants.js
@@ -1,0 +1,43 @@
+const notificationTypes = Object.freeze({
+  Follow: 'Follow',
+  Repost: {
+    base: 'Repost',
+    track: 'RepostTrack',
+    album: 'RepostAlbum',
+    playlist: 'RepostPlaylist'
+  },
+  Favorite: {
+    base: 'Favorite',
+    track: 'FavoriteTrack',
+    album: 'FavoriteAlbum',
+    playlist: 'FavoritePlaylist'
+  },
+  Create: {
+    base: 'Create',
+    track: 'CreateTrack',
+    album: 'CreateAlbum',
+    playlist: 'CreatePlaylist'
+  },
+  MilestoneFollow: 'MilestoneFollow',
+  MilestoneRepost: 'MilestoneRepost',
+  MilestoneFavorite: 'MilestoneFavorite',
+  MilestoneListen: 'MilestoneListen',
+  Announcement: 'Announcement'
+})
+
+const actionEntityTypes = Object.freeze({
+  User: 'User',
+  Track: 'Track',
+  Album: 'Album',
+  Playlist: 'Playlist'
+})
+
+const dayInHours = 24
+const weekInHours = 168
+
+module.exports = {
+  notificationTypes,
+  actionEntityTypes,
+  dayInHours,
+  weekInHours
+}
